Merged the functionality of unit_stats_writer into full_table_writer. If the save_iteration_outputs_for_units flag in the database is set, then full_table_writer appends an extra section to its output, saying how many times units were infected, vaccinated, etc. during the run.  #381 The database is now opened readonly as is closed as soon as the simulator is finished reading from it.
Note that the function import_naadsm_xml now has a keyword argument to set save_iteration_outputs_for_units, since that is on by default but I want it off when running the automated test suite. Also, I set that flag to off manually in the one test that exists only as a sqlite3 file.